% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20072R.R
\name{Ftest}
\alias{Ftest}
\title{F Test to Compare Two Variances}
\usage{
Ftest(x, y)
}
\arguments{
\item{x}{A set of data}

\item{y}{A set of data that is different from \code{x}}
}
\value{
p-value of F Test to Compare Two Variances
}
\description{
F Test to Compare Two Variances and compute the P value
}
\examples{
\dontrun{
sigma1 <- 1 
sigma2 <- 1.5
m<-1000
n<-c(10,50,100)
power<-numeric(m)
for(j in 1:length(n)){
     for(i in 1:m){
      x<-rnorm(n[j],0,sigma1)
       y<-rnorm(n[j],0,sigma2)
       p.Ftest<-Ftest(x,y)
     }
    power[j]<-p.Ftest
   }
   power
}
}
