% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20072R.R
\name{maxout}
\alias{maxout}
\title{Number of extreme points}
\usage{
maxout(x, y)
}
\arguments{
\item{x}{A set of data}

\item{y}{A set of data compared to \code{x}}
}
\value{
the number of extreme points
}
\description{
Counts the number of extreme points of each sample relative to the range of the other sample.
}
\examples{
\dontrun{
set.seed(124)
x<-rnorm(20,4,2)
y<-rnorm(10,4,2)
test<-maxout(x,y)
}
}
