% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MetropolisC}
\alias{MetropolisC}
\title{Generate the standard Laplace distribution with a normal increment using Rcpp}
\usage{
MetropolisC(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard error}

\item{x0}{the intial value}

\item{N}{the simulation times}
}
\value{
a random samples \code{n}
}
\description{
Generate the standard Laplace distribution with a normal increment using Rcpp
}
\examples{
\dontrun{
sigma<-1
x0<-25
N<-3000
MC<-MetropolisC(sigma,x0,N)
}
}
