% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20072R.R
\name{MetropolisR}
\alias{MetropolisR}
\title{Generate the standard Laplace distribution with a normal increment}
\usage{
MetropolisR(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard error}

\item{x0}{the intial value}

\item{N}{the simulation times}
}
\value{
a random sample of size \code{n}
}
\description{
Generate the standard Laplace distribution with a normal increment
}
\examples{
\dontrun{
N<-3000
sigma<-1
x0<-25
MR1<-MetropolisR(sigma,x0,N)
}
}
