% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20072R.R
\name{Mnorm}
\alias{Mnorm}
\title{Mixed normal distribution}
\usage{
Mnorm(p, N, mu0, sigma0)
}
\arguments{
\item{p}{The weight vector}

\item{N}{The number of random numbers}

\item{mu0}{The mean vector}

\item{sigma0}{The standard error vector}
}
\value{
a random sample of size \code{n}
}
\description{
Generate a mixed normal distribution by weight
}
\examples{
\dontrun{
p<-c(0.2,0.2,0.1,0.5)
sigma0<-c(1,2,4,5)
mu0<-c(1,3,4,2)
N<-20
Mnorm(p,N,mu0,sigma0)
}
}
