% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20072R.R
\name{observed}
\alias{observed}
\title{Observed data likelihood}
\usage{
observed(x)
}
\arguments{
\item{x}{initial value of p and q}
}
\value{
data likelihood
}
\description{
Observed data likelihood
}
\examples{
\dontrun{
theta<- optim(c(0.35,0.2),fn=observed)$par
}
}
