% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20072R.R
\name{rSim}
\alias{rSim}
\title{Simulations in VAR systems}
\usage{
rSim(coeff, errors)
}
\arguments{
\item{coeff}{the coefficient matrix}

\item{errors}{error terms}
}
\value{
Data matrix
}
\description{
Simulations in VAR systems(Explicit loop)
}
\examples{
\dontrun{
a<-matrix(c(0.5,0.1,0.1,0.5),nrow=2)
u<-matrix(rnorm(10000),ncol=2)
rData<-rSim(a,u)
}
}
