% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20072R.R
\name{Rubin}
\alias{Rubin}
\title{G-R statistic}
\usage{
Rubin(psi)
}
\arguments{
\item{psi}{A matrix with n rows and 2 columns}
}
\value{
G-R statistic
}
\description{
Compute G-R statistic
}
\examples{
\dontrun{
sigma1<-0.05
N<-20000
b<-1000
K<-4
x0<-c(-8,-4,4,8)
X<-matrix(0,K,N)
for(i in 1:K)
 X[i,]<-Metropolis(sigma1,x0[i],N)
psi1<-t(apply(X,1,cumsum))
for(i in 1:nrow(psi1))
 psi1[i,]<-psi1[i,]/(1:ncol(psi1))
print(Rubin(psi1))
}
}
