% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20072R.R
\name{skewness}
\alias{skewness}
\title{Compute skewness coeff}
\usage{
skewness(x)
}
\arguments{
\item{x}{A set of data}
}
\value{
skewness coeff
}
\description{
compute the skewness coeff of a set of data
}
\examples{
\dontrun{p.reject <- numeric(length(n)) 
 m <- 10000
 for (i in 1:length(n)) { sktests <- numeric(m)
 for (j in 1:m) { x <- rnorm(n[i])
 sktests[j] <- as.integer(abs(skewness(x)) >= cv[i] ) 
 }
 p.reject[i] <- mean(sktests)}
}
}
