\name{StatComp20072-package}
\alias{StatComp20072-package}
\alias{StatComp20072}
\docType{package}
\title{
  some functions using in "Statistical Computing"
}
\description{Many statistical analyses and numerical calculations can be performed.
}

\author{Fangjuan Ye

Maintainer: Fangjuan Ye <yfj0530@mail.ustc.edu.cn>
}

\keyword{ package }

